Update Macrotrends.py
First, thanks for providing the scraper. I thought that was extremely useful for people in finance. To pay attributes, I've made some changes to the script by adding some functions for convenience. Here are some changes that I've made: 

Instead of exporting .csv, I export to excel file. (trivial change)

Multiple URLs and Sheets: This change was made in the main() function. The script was modified to scrape data from multiple URLs (Balance Sheet, Cash Flow Statement, Key Metrics, and Income Statement) and write the data to separate sheets in the same Excel file.

Ticker Placeholder: A placeholder for ticker symbols was added to the main() function. The script now accepts a list of ticker symbols and iterates over them to generate the URLs for scraping and to name the Excel files.

Company Name Placeholder: This change was also made in the main() function. This is to retrieve the quarterly data.

Row Index Formatting: The clean_row_index(df) function was added to clean up the row index.

Numeric Data Type Conversion: The convert_to_numeric(df) function was added to convert the numbers to numeric data type.

Text Alignment: The align_text(filename, sheet_names) function was added to align the text to the left in the first column.

Number Formatting: The format_numbers(filename, sheet_names) function was added to format the numbers to have two decimal places.

Column Width Autofitting: The autofit_column_widths(filename, sheet_names) function was added to autofit the column widths.